refactor(suite-desktop): unify entrypoint with web
